put config path finding and exceptions into own modules
